crud functions completed
The main changes in these commits have to due with the core functions of
the app being fully implemented. Creating, reading, updating, and delete
now work as expected. With these core functions in place I will work on
modularizing them so they make sense. The implementation now is a bit
hacky and could be cleaned up to make more sense.

Any of the creating updating deleting will call ajax js requests to routes
on the server that do the requested action.

Implemented extremely basic feedback using primitive js function for
alerting the user. Sets the foundation for later moving to a more inline
system that doesn't use these older features.